[#21] database: Refactor database
Splits the 'database.go' into three files:

* database.go - responsible for manipulating databases
* offers.go - responsible for manipulating offers
* searches.go - responsible for manipulating searches

Updates docstrings in the 'database' and adds if absent.

Minor fixes to 'telegrambot'.
